If I understand translations correctly, this step is not needed anymore:
Back in the days we had plonelink and ploneimage plugins and they used
translations of advlink and advimage plugins that come with TinyMCE

Now we have only one plugin (for both images and links) and it is suppprted
in it's entirety by plone.tinymce.po language files.

Hence the steps described here are not needed any more.